rename: modal 컴포넌트, kakaomap 컴포넌트 파일 폴더 분리
